Rename modules
Renamed:
MDNX_API -> CR_MDNX_API

Created:
HIDIVE_MDNX_API
